+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+- Added the multi-year and multi-crop selection feature to the crop insurance page [#466](https://github.com/policy-design-lab/pdl-frontend/issues/466)
+
+### Changed
+- Updated the site to reusable components such as Export to CSV button [#468](https://github.com/policy-design-lab/pdl-frontend/issues/468)
+
 ## [1.18.0] - 2026-08-27
 
 ### Added